refactor: split Stripe billing module
